requirements.txt
# --- Core LangChain stack ---
langchain==0.3.25
langchain-core==0.3.62
langchain-community==0.3.24
langchain-openai==0.3.18
langsmith==0.3.42

# --- Server & Web Frameworks ---
fastapi==0.115.9
uvicorn==0.22.0
streamlit==1.45.1

# --- AI & NLP ---
openai==1.82.0
ollama
sentence-transformers==4.1.0
transformers==4.52.3
tiktoken==0.9.0

# --- Vector DB & Search ---
chromadb==1.0.11
faiss-cpu==1.11.0

# --- Utilities & Helpers ---
python-dotenv==1.1.0
httpx==0.28.1
httpx-sse==0.4.0
requests==2.32.3
requests-cache==1.2.1
orjson==3.10.18
protobuf==6.31.1
pydantic==2.11.5
pydantic-settings==2.9.1

# --- Data Science ---
pandas==2.2.3
numpy==2.2.2
scikit-learn==1.6.1
matplotlib==3.10.0
seaborn==0.13.2
scipy==1.15.1
statsmodels==0.14.4

# --- Dev & Notebook ---
jupyter==1.1.1
notebook==7.3.2
ipywidgets==8.1.5

# --- Logging & Debug ---
rich==14.0.0
tenacity==9.1.2
loguru==0.7.2

# --- Serialization ---
PyYAML==6.0.2
jsonschema==4.23.0

# --- Caching & Tokenizer Utils ---
langchainhub==0.1.21
arxiv==2.2.0
wikipedia==1.4.0
Wikipedia-API==0.8.1